// src/app/features/KetQuaHocTap/Item/type.ts
export interface BieuDoProp {
  _id: string;
  maSvHk: string;
  thuTuHocKy: number;
  trungBinhHocKy: number;
  trungBinhHocKyThang4: number;
  trungBinhTichLuyToanKhoa: null;
  trungBinhTichLuyToanKhoaThang4: null;
  trungBinhHocBongHocKy: null;
  trungBinhHocBongHocKyThang4: null;
  tongSoTinChiDangKyToanKhoa: number;
  tongSoTinChiHocKy: number;
  tongSoTinChiTichLuyHocKy: number;
  tongSoTinChiTichLuyToanKhoa: number;
  tongSoTinChiNoHocKy: number;
  tongSoTinChiNoToanKhoa: number;
  tongSoTinChiHocLai: number;
  tongSoTinChiHocCaiThien: number;
  tongSoTinChiThiLai: number;
  tongSoTinChiThiCaiThien: number;
  tongSoTinChiDangKyHocKy: number;
  trinhDo: string;
  hocLuc: string;
  createdAt: Date;
  updatedAt: Date;
  sinhVienSsoId: string;
  maHocKy: string;
}